+++ src/main/java/be/baur/sdt/SDT.java
@@ -3,7 +3,7 @@ package be.baur.sdt;
 import java.io.IOException;
 import java.io.Writer;
 
-import be.baur.sdt.serialization.Parser;
+import be.baur.sdt.serialization.TransformParser;
 import be.baur.sdt.serialization.SDTParser;
 
 /**
@@ -49,7 +49,7 @@ public final class SDT {
 	 * 
 	 * @return an {@link SDTParser}
 	 */
-	public static Parser parser() {
+	public static TransformParser parser() {
 		return new SDTParser();
 	}