+++ src/main/java/be/baur/sdt/TransformException.java
@@ -1,7 +1,7 @@
 package be.baur.sdt;
 
 import be.baur.sda.Node;
-import be.baur.sda.serialization.NodeProcessingException;
+import be.baur.sda.NodeProcessingException;
 
 /**
  * An {@code TransformException} may be thrown during the execution of a